move generateObfuscatedMockRac to script file
